requirements.txt
streamlit
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
requests
bcrypt

# Development dependencies
pytest
pytest-cov
flake8
black